analyzer: Move PythonInspector to a separate file
Move the `PythonInspector` object to a separate file as a preparation
for adding deserialization classes.
